re-organized file structure to separate new scheme from previous; added attempt at using NeuroscopeSortingExtractor, including data
