--- buzsaki_lab_to_nwb/__init__.py
@@ -1,1 +1,0 @@
-from .conversion_module import *
