--- buzsaki_lab_to_nwb/original_code/extensions/test_type.py
@@ -1,4 +1,0 @@
-import pynwb
-
-pynwb.spec.NWBAttributeSpec(name='my_name', dtype='sdfsd', doc='my doc')
-
