--- buzsaki_lab_to_nwb/original_code/general.namespace.yaml
@@ -1,6 +1,0 @@
-namespaces:
-- doc: general extensions
-  name: general
-  schema:
-  - namespace: core
-  - source: general.extensions.yaml
