+++ buzsaki_lab_to_nwb/schema-v2.0b/metafile.yml
@@ -1,0 +1,19 @@
+NWBFile:
+  data_collection: ''
+  experiment_description: ADDME
+  experimenter: Experimenter
+  identifier: ABC123
+  institution: NYU
+  keywords:
+  - ''
+  lab: Buzsaki lab
+  notes: ''
+  pharmacology: ''
+  protocol: ''
+  related_publications: ''
+  session_description: ''
+  session_id: ''
+  session_start_time: 2018-08-09 10:00:00
+  slices: ''
+  stimulus_notes: ''
+  surgery: ''
