+++ buzsaki_lab_to_nwb/schema-v2.0b/original_code/extensions/test_type.py
@@ -1,0 +1,4 @@
+import pynwb
+
+pynwb.spec.NWBAttributeSpec(name='my_name', dtype='sdfsd', doc='my doc')
+
