+++ buzsaki_lab_to_nwb/schema-v2.2.5/original_code/general.extensions.yaml
@@ -1,0 +1,42 @@
+groups:
+- datasets:
+  - doc: global id for neuron
+    dtype: int
+    name: cell_index
+    quantity: '?'
+    shape:
+    - null
+    - 1
+  - attributes:
+    - doc: values that the indices are indexing
+      dtype: text
+      name: values
+      shape:
+      - null
+      - 1
+    doc: list of indices for values
+    dtype: int
+    name: indices
+    shape:
+    - null
+    - 1
+  doc: Categorical Cell Info
+  neurodata_type_def: CatCellInfo
+  neurodata_type_inc: NWBDataInterface
+- datasets:
+  - attributes:
+    - doc: values that the indices are indexing
+      dtype: text
+      name: values
+      shape:
+      - null
+      - 1
+    doc: timeseries of indicies for values
+    dtype: int
+    name: data
+    shape:
+    - null
+    - 1
+  doc: Categorical data through time
+  neurodata_type_def: CatTimeSeries
+  neurodata_type_inc: TimeSeries
