+++ buzsaki_lab_to_nwb/schema-v2.2.5/original_code/general.namespace.yaml
@@ -1,0 +1,6 @@
+namespaces:
+- doc: general extensions
+  name: general
+  schema:
+  - namespace: core
+  - source: general.extensions.yaml
